Rename stimulus controller to b4
