+++ src/UDS.Net.Forms/Pages/UDS4/B4.cshtml
@@ -10,7 +10,7 @@
         INSTRUCTIONS: For information on the required online CDR training, see <guidebook kind="@Model.Visit.PACKET"></guidebook>, Form B4. This form is to be completed by the clinician or other trained health professional, based on co-participant report and behavioral and neurological exam of the participant. In the extremely rare instances when no co-participant is available, the clinician or other trained health professional must complete this form using all other available information and his/her best clinical judgment. Score only as decline from previous level due to <span class="underline">cognitive loss</span>, not impairment due to other factors, such as physical disability. For further information, see <guidebook kind="@Model.Visit.PACKET"></guidebook>, Form B4.
     </p>
 </div>
-<div data-controller="autocalculate">
+<div data-controller="b4">
     <div class="mt-10 space-y-8 border-b border-gray-900/10 pb-12 sm:space-y-0 sm:divide-y sm:divide-gray-900/10 sm:border-t sm:pb-0">
         <div class="sm:grid sm:grid-cols-3 sm:items-start sm:gap-4 sm:py-6">
             <h2>SECTION 1: CDR DEMENTIA STAGING INSTRUMENT<sup>1</sup></h2>
@@ -44,7 +44,7 @@
                     <td class="mt-1 p-2 text-sm text-gray-500">Severe memory loss; only highly learned material retained; new material rapidly lost</td>
                     <td class="mt-1 p-2 text-sm text-gray-500">Severe memory loss; only fragments remain</td>
                 </tr>
-                <tr data-autocalculate-target="MEMORY">
+                <tr data-b4-target="MEMORY">
                     <table-cell-radio-buttons id="@Html.IdFor(m => m.B4.MEMORY)" for="@Model.B4.MEMORY" items="Model.ImpairmentListItems"></table-cell-radio-buttons>
                 </tr>
                 <tr class="divide-y divide-gray-200">
@@ -59,7 +59,7 @@
                     <td class="mt-1 p-2 text-sm text-gray-500">Severe difficulty with time relationships; usually disoriented to time, often to place</td>
                     <td class="mt-1 p-2 text-sm text-gray-500">Oriented to person only</td>
                 </tr>
-                <tr data-autocalculate-target="ORIENT">
+                <tr data-b4-target="ORIENT">
                     <table-cell-radio-buttons id="@Html.IdFor(m => m.B4.ORIENT)" for="@Model.B4.ORIENT" items="Model.ImpairmentListItems"></table-cell-radio-buttons>
                 </tr>
                 <tr class="divide-y divide-gray-200">
@@ -74,7 +74,7 @@
                     <td class="mt-1 p-2 text-sm text-gray-500">Severely impaired in handling problems, similarities, and differences; social judgment usually impaired</td>
                     <td class="mt-1 p-2 text-sm text-gray-500">Unable to make judgments or solve problems</td>
                 </tr>
-                <tr data-autocalculate-target="JUDGMENT">
+                <tr data-b4-target="JUDGMENT">
                     <table-cell-radio-buttons id="@Html.IdFor(m => m.B4.JUDGMENT)" for="@Model.B4.JUDGMENT" items="Model.ImpairmentListItems"></table-cell-radio-buttons>
                 </tr>
                 <tr class="divide-y divide-gray-200">
@@ -89,7 +89,7 @@
                     <td class="mt-1 p-2 text-sm text-gray-500">No pretense of independent function outside the home; appears well enough to be taken to functions outside the family home</td>
                     <td class="mt-1 p-2 text-sm text-gray-500">No pretense of independent function outside the home; appears too ill to be taken to functions outside the family home</td>
                 </tr>
-                <tr data-autocalculate-target="COMMUN">
+                <tr data-b4-target="COMMUN">
                     <table-cell-radio-buttons id="@Html.IdFor(m => m.B4.COMMUN)" for="@Model.B4.COMMUN" items="Model.ImpairmentListItems"></table-cell-radio-buttons>
                 </tr>
                 <tr class="divide-y divide-gray-200">
@@ -104,7 +104,7 @@
                     <td class="mt-1 p-2 text-sm text-gray-500">Only simple chores preserved; very restricted interests, poorly maintained</td>
                     <td class="mt-1 p-2 text-sm text-gray-500">No significant function in the home</td>
                 </tr>
-                <tr data-autocalculate-target="HOMEHOBB">
+                <tr data-b4-target="HOMEHOBB">
                     <table-cell-radio-buttons id="@Html.IdFor(m => m.B4.HOMEHOBB)" for="@Model.B4.HOMEHOBB" items="Model.ImpairmentListItems"></table-cell-radio-buttons>
                 </tr>
                 <tr class="divide-y divide-gray-200">
@@ -118,7 +118,7 @@
                     <td class="mt-1 p-2 text-sm text-gray-500">Requires assistance in dressing, hygiene, keeping of personal effects</td>
                     <td class="mt-1 p-2 text-sm text-gray-500">Requires much help with personal care; frequent incontinence</td>
                 </tr>
-                <tr data-autocalculate-target="PERSCARE">
+                <tr data-b4-target="PERSCARE">
                     <!--// Only 4 options for Personal Care (PERSCARE) so the first two cells need to be merged //-->
                     @for (int i = 0; i < Model.ImpairmentSimpleListItems.Count; i++)
                     {
@@ -141,7 +141,7 @@
             <label asp-for="B4.CDRSUM"><span class="counter"></span> @Html.DisplayNameFor(m => m.B4.CDRSUM)</label>
             <div class="mt-4 sm:col-span-2 sm:mt-0">
                 <p class="text-sm text-gray-500" asp-description-for="@Model.B4.CDRSUM"></p>
-                <input asp-for="@Model.B4.CDRSUM" data-autocalculate-target="CDRSUM" readonly/>
+                <input asp-for="@Model.B4.CDRSUM" data-b4-target="CDRSUM" readonly/>
                 <span class="mt-2 text-sm text-red-600" asp-validation-for="B4.CDRSUM"></span>
             </div>
         </div>
@@ -150,7 +150,7 @@
             <label asp-for="B4.CDRGLOB"><span class="counter"></span> @Html.DisplayNameFor(m => m.B4.CDRGLOB)</label>
             <div class="mt-4 sm:col-span-2 sm:mt-0">
                 <p class="text-sm text-gray-500" asp-description-for="@Model.B4.CDRGLOB"></p>
-                <radio-button-group id="@Html.IdFor(m => m.B4.CDRGLOB)" for="@Model.B4.CDRGLOB" items="Model.ImpairmentListItems" data-autocalculate-target="CDRGLOB">
+                <radio-button-group id="@Html.IdFor(m => m.B4.CDRGLOB)" for="@Model.B4.CDRGLOB" items="Model.ImpairmentListItems" data-b4-target="CDRGLOB">
                 </radio-button-group>
                 <span class="mt-2 text-sm text-red-600" asp-validation-for="B4.CDRGLOB"></span>
             </div>
