+++ src/UDS.Net.Forms/wwwroot/js/stimulus_controllers.js
@@ -8,7 +8,7 @@ import selectendpoint_controller from "./js_controllers/selectendpoint_controlle
 import checkboxSelectAll from "./js_controllers/checkboxSelectAll_controller.js"
 import rxNormDisplayNames from "./js_controllers/rxNormDisplayNames_controller.js"
 import autocomplete from "./js_controllers/autocomplete_controller.js"
-import autocalculate from "./js_controllers/autocalculate_controller.js"
+import b4 from "./js_controllers/b4_controller.js"
 import b6 from "./js_controllers/b6_controller.js"
 
 window.Stimulus = Application.start()
@@ -22,5 +22,5 @@ Stimulus.register("selectendpoint", selectendpoint_controller)
 Stimulus.register("checkboxSelectAll", checkboxSelectAll)
 Stimulus.register("rxNormDisplayNames", rxNormDisplayNames)
 Stimulus.register("autocomplete", autocomplete)
-Stimulus.register("autocalculate", autocalculate)
+Stimulus.register("b4", b4)
 Stimulus.register("b6", b6)
